nutrisignal_app/backend/requirements.txt
fastapi
uvicorn
tensorflow-macos
tensorflow-metal
numpy
pillow
requests
python-multipart